Revert: Switch content of purification and questions
